Added left and right bounds; ( command
All the different shapes which previously existed as different
constructors of tetramino are now each their own class.

I did it this way so I could dispatch methods on the (namely, the
get-x-offset series of methods).

Also added the ( command, or turn -90 degrees. Naively implemented as a
series of 3 90 degree turns, but performance isn’t a concern so
whatever’s easiest.