Drop command, set down blocks, refactoring
New symbol `position-next` is simply used for figuring out if next
movement will have a collision.

Two generics created for tetramino, `uppers` and `lowers` for use in
printing.

`<empty-tetr>` constructor added so we can represent a state where
there is no active tetramino.

`intersect-with-active-at?` lets us put in an arbitrary position for
active to be checked for intersections (for example, we can use
`position-next` instead of `position`).

`print-matrix-with-active` now refers to a helper word,
`get-matrix-with-active`, as this functionality is also used in
`set-down-active-tetr`.

Also added V: drop command.