Down-movement collision detection
So far only works for V-command.

This was my priority, and my code got messy. But it works! I’ll clean
it up later (I promise!).

Some messy parts:

I abandoned the “position-next” scheme, but haven’t yet removed that
code and reverted it back to assuming position is only position. (Or
should I keep it where functions can have a position input? This keeps
code more generic).

`next-has-collision?` is way too long.

I’m pretty sure I’m mixing up 0-index and 1-index logic.

`move-down`’s logic for bottom-collision has been moved to
`tetra-collides-with-grid?` and is called by `next-has-collision?`
(which should be renamed.

`next-has-collision` simply looks for any collisions. First it gets a
y-coordinate and a list of x-coordinates that is the bottom row of a
tetramino, and it calls `tetra-collides-with-grid?` on them.

`tetra-collides-with-grid?` checks for height (does it collide with
“empty” grid?), then checks if any of a mapped
`tetra-collides-with-grid-at?` on the y and x-coords. It then has to
drop the extra y used for the mapping.

`tetra-collides-with-grid-at?` checks both the bottom row, and the row
above it.

I hope the rest is self-explanatory.